requirements.txt
flask==2.0.2
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
pyjwt==2.3.0
pytest==6.1.1
